# First iteration of performance tweaks