--- a/.travis.yml
+++ b/.travis.yml
@@ -2,5 +2,7 @@ language: node_js
 sudo: false
 node_js:
   - 6
+  - 5
+  - 4
 after_script:
   - npm run coveralls