--- a/.travis.yml
+++ b/.travis.yml
@@ -1,8 +1,16 @@
 language: node_js
-sudo: false
 node_js:
+  - node
   - 6
-  - 5
   - 4
-after_script:
-  - npm run coveralls
+cache:
+  directories:
+    - node_modules
+before_script:
+  - npm i
+script:
+  - npm test
+after_success:
+- './node_modules/.bin/nyc report --reporter=text-lcov | ./node_modules/.bin/coveralls'
+notifications:
+  email: false