--- a/lib/index.js
+++ b/lib/index.js
@@ -1,26 +1,55 @@
 const vm = require('vm')
 const merge = require('lodash.merge')
 const cloneDeep = require('lodash.clonedeep')
-const parseAndReplace = require('./expression_parser')
+const placeholders = require('./placeholders')
 
-let delimiters, unescapeDelimiters, delimiterRegex, unescapeDelimiterRegex, conditionals, loops, scopes, options
+let delimitersSettings = []
+let conditionals, loops, scopes, options
 
-module.exports = function PostHTMLExpressions (_options = {}) {
-  options = _options
-  // set up delimiter options and detection
-  delimiters = options.delimiters || ['{{', '}}']
-  unescapeDelimiters = options.unescapeDelimiters || ['{{{', '}}}']
-  delimiterRegex = new RegExp(`.*${delimiters[0]}(.*)${delimiters[1]}.*`, 'g')
-  unescapeDelimiterRegex = new RegExp(`.*${unescapeDelimiters[0]}(.*)${unescapeDelimiters[1]}.*`, 'g')
+function escapeRegexpString (input) {
+  return input.replace(/[|\\{}()[\]^$+*?.]/g, '\\$&')
+}
 
-  // identification for delimiter options, for the parser
-  delimiters.push('escaped')
-  unescapeDelimiters.push('unescaped')
+module.exports = function PostHTMLExpressions (_options) {
+  // set default options
+  options = Object.assign({
+    delimiters: ['{{', '}}'],
+    unescapeDelimiters: ['{{{', '}}}'],
+    conditionalTags: ['if', 'elseif', 'else'],
+    loopTags: ['each'],
+    scopeTags: ['scope'],
+    locals: {}
+  }, _options)
 
-  // conditional and loop options
-  conditionals = options.conditionalTags || ['if', 'elseif', 'else']
-  scopes = options.scopeTags || ['scope']
-  loops = options.loopTags || ['each']
+  loops = options.loopTags
+  scopes = options.scopeTags
+  conditionals = options.conditionalTags
+
+  // make a RegExp's to search for placeholders
+  let before = escapeRegexpString(options.delimiters[0])
+  let after = escapeRegexpString(options.delimiters[1])
+  const delimitersRegexp = new RegExp(`${before}(.+?)${after}`, 'g')
+
+  before = escapeRegexpString(options.unescapeDelimiters[0])
+  after = escapeRegexpString(options.unescapeDelimiters[1])
+  const unescapeDelimitersRegexp = new RegExp(`${before}(.+?)${after}`, 'g')
+
+  // make array of delimiters
+  const delimiters = [
+    { text: options.delimiters, regexp: delimitersRegexp, escape: true },
+    { text: options.unescapeDelimiters, regexp: unescapeDelimitersRegexp, escape: false }
+  ]
+
+  // we arrange delimiter search order by length, since it's possible that one
+  // delimiter could 'contain' another delimiter, like '{{' and '{{{'. But if
+  // you sort by length, the longer one will always match first.
+  if (options.delimiters.join().length > options.unescapeDelimiters.join().length) {
+    delimitersSettings[0] = delimiters[0]
+    delimitersSettings[1] = delimiters[1]
+  } else {
+    delimitersSettings[0] = delimiters[1]
+    delimitersSettings[1] = delimiters[0]
+  }
 
   // kick off the parsing
   return walk.bind(null, options)
@@ -40,12 +69,8 @@ function walk (opts, nodes) {
     if (skip) { skip--; return m }
 
     // if we have a string, match and replace it
-    if (typeof node !== 'object') {
-      // if there are any matches, we parse and replace the results
-      if (node.match(delimiterRegex) || node.match(unescapeDelimiterRegex)) {
-        // TODO: this could be optimized by starting at the regex match index
-        node = parseAndReplace(ctx, [delimiters, unescapeDelimiters], node)
-      }
+    if (typeof node === 'string') {
+      node = placeholders(node, ctx, delimitersSettings)
       m.push(node)
       return m
     }
@@ -53,10 +78,7 @@ function walk (opts, nodes) {
     // if not, we have an object, so we need to run the attributes and contents
     if (node.attrs) {
       for (let key in node.attrs) {
-        const val = node.attrs[key]
-        if (val.match(delimiterRegex)) {
-          node.attrs[key] = parseAndReplace(ctx, [delimiters, unescapeDelimiters], val)
-        }
+        node.attrs[key] = placeholders(node.attrs[key], ctx, delimitersSettings)
       }
     }