--- a/lib/index.js
+++ b/lib/index.js
@@ -223,60 +223,29 @@ function getNextTag (nodes, i, nodeCount) {
 
 /**
  * Given a "loop" parameter from an "each" tag, parses out the param names and
- * expression to be looped through using a mini text parser.
+ * expression to be looped.
  */
 function parseLoopStatement (input) {
-  let current = 0
-  let char = input[current]
+  // try to find ` in ` keyword
+  const inKeywordIndex = input.search(/\sin\s/)
 
-  // parse through keys `each **foo, bar** in x`, which is everything before
-  // the word "in"
-  const keys = []
-  let key = ''
-  while (!`${char}${lookahead(3)}`.match(/\sin\s/)) {
-    key += char
-    next()
-
-    // if we hit a comma, we're on to the next key
-    if (char === ',') {
-      keys.push(key.trim())
-      key = ''
-      next()
-    }
-
-    // if we reach the end of the string without getting "in", it's an error
-    if (typeof char === 'undefined') {
-      throw new Error("Loop statement lacking 'in' keyword")
-    }
-  }
-  keys.push(key.trim())
-
-  // Bypass the word " in", and ensure there's a space after
-  next(4)
-
-  // the rest of the string is evaluated as the array/object to loop
-  let expression = ''
-  while (current < input.length) {
-    expression += char
-    next()
+  // if we reach the end of the string without getting "in", it's an error
+  if (inKeywordIndex === -1) {
+    throw new Error("Loop statement lacking 'in' keyword")
   }
 
-  return {keys, expression}
+  // expression is always after `in` keyword
+  const expression = input.substr(inKeywordIndex + 4)
 
-  // Utility: Move to the next character in the parse
-  function next (n = 1) {
-    for (let i = 0; i < n; i++) { char = input[++current] }
+  // keys is always before `in` keyword
+  const keys = input.substr(0, inKeywordIndex).split(',')
+  for (let i = 0; i < keys.length; i++) {
+    keys[i] = keys[i].trim()
   }
 
-  // Utility: looks ahead n characters and returns the result
-  function lookahead (n) {
-    let counter = current
-    const target = current + n
-    let res = ''
-    while (counter < target) {
-      res += input[++counter]
-    }
-    return res
+  return {
+    keys,
+    expression
   }
 }