--- a/lib/index.js
+++ b/lib/index.js
@@ -95,22 +95,24 @@ function walk (opts, nodes) {
         throw new Error(`the "${conditionals[0]}" tag must have a "condition" attribute`)
       }
 
-      // build the expression object, which we will turn into js and eval later
-      const ast = [{
-        statement: 'if',
-        condition: node.attrs.condition,
-        content: node.content
-      }]
+      // сalculate the first path of condition expression
+      let expressionIndex = 1
+      let expression = `if (${node.attrs.condition}) { 0 } `
+      const branches = [node.content]
 
       // move through the nodes and collect all others that are part of the same
       // conditional statement
-      let [current, nextTag] = getNextTag(nodes, ++i)
-      while (conditionals.slice(-2).indexOf(nextTag.tag) > -1) {
-        const obj = { statement: nextTag.tag, content: nextTag.content }
+      let computedNextTag = getNextTag(nodes, ++i)
+      let current = computedNextTag[0]
+      let nextTag = computedNextTag[1]
+
+      while (conditionals.slice(1).indexOf(nextTag.tag) > -1) {
+        let statement = nextTag.tag
+        let condition = ''
 
         // ensure the "else" tag is represented in our little AST as 'else',
         // even if a custom tag was used
-        if (nextTag.tag === conditionals[2]) obj.statement = 'else'
+        if (nextTag.tag === conditionals[2]) statement = 'else'
 
         // add the condition if it's an else if
         if (nextTag.tag === conditionals[1]) {
@@ -118,32 +120,31 @@ function walk (opts, nodes) {
           if (!(nextTag.attrs && nextTag.attrs.condition)) {
             throw new Error(`the "${conditionals[1]}" tag must have a "condition" attribute`)
           }
-          obj.condition = nextTag.attrs.condition
+          condition = nextTag.attrs.condition
 
           // while we're here, expand "elseif" to "else if"
-          obj.statement = 'else if'
+          statement = 'else if'
         }
-        ast.push(obj)
+        branches.push(nextTag.content)
 
-        ;[current, nextTag] = getNextTag(nodes, ++current)
+        // calculate next part of condition expression
+        expression += statement + (condition ? ` (${condition})` : '') + ` { ${expressionIndex++} } `
+
+        computedNextTag = getNextTag(nodes, ++current)
+        current = computedNextTag[0]
+        nextTag = computedNextTag[1]
       }
 
-      // format into an expression
-      const expression = ast.reduce((m2, e, i) => {
-        m2 += e.statement
-        if (e.condition) m2 += ` (${e.condition})`
-        m2 += ` { ${i} } `
-        return m2
-      }, '')
-
-      // evaluate the expression, get the winning node
-      const expResult = ast[vm.runInContext(expression, ctx)]
+      // evaluate the expression, get the winning condition branch
+      const branch = branches[vm.runInContext(expression, ctx)]
 
       // remove all of the conditional tags from the tree
       // we subtract 1 from i as it's incremented from the initial if statement
       // in order to get the next node
       skip = current - i
-      if (expResult) m.push(...walk(opts, expResult.content))
+
+      // recursive evaluate of condition branch
+      if (branch) Array.prototype.push.apply(m, walk(options, branch))
       return m
     }