--- a/lib/index.js
+++ b/lib/index.js
@@ -1,3 +1,5 @@
+'use strict'
+
 const vm = require('vm')
 const merge = require('lodash.merge')
 const cloneDeep = require('lodash.clonedeep')