--- a/lib/index.js
+++ b/lib/index.js
@@ -6,13 +6,13 @@ const cloneDeep = require('lodash.clonedeep')
 const placeholders = require('./placeholders')
 
 let delimitersSettings = []
-let conditionals, loops, scopes, options
+let conditionals, loops, scopes, data
 
 function escapeRegexpString (input) {
   return input.replace(/[|\\{}()[\]^$+*?.]/g, '\\$&')
 }
 
-module.exports = function PostHTMLExpressions (_options) {
+module.exports = function PostHTMLExpressions (options) {
   // set default options
   options = Object.assign({
     delimiters: ['{{', '}}'],
@@ -21,8 +21,12 @@ module.exports = function PostHTMLExpressions (_options) {
     loopTags: ['each'],
     scopeTags: ['scope'],
     locals: {}
-  }, _options)
+  }, options)
 
+  // make data for expressions
+  data = { locals: options.locals }
+
+  // set tags
   loops = options.loopTags
   scopes = options.scopeTags
   conditionals = options.conditionalTags
@@ -54,7 +58,7 @@ module.exports = function PostHTMLExpressions (_options) {
   }
 
   // kick off the parsing
-  return walk.bind(null, options)
+  return walk.bind(null, data)
 }
 
 function walk (opts, nodes) {
@@ -146,7 +150,7 @@ function walk (opts, nodes) {
       skip = current - i
 
       // recursive evaluate of condition branch
-      if (branch) Array.prototype.push.apply(m, walk(options, branch))
+      if (branch) Array.prototype.push.apply(m, walk(opts, branch))
       return m
     }
 
@@ -272,7 +276,7 @@ function executeLoop (loopParams, p1, p2, node) {
  */
 function executeScoped (scopedLocals, node) {
   // merge nondestructively into existing locals
-  const scopedOptions = merge(cloneDeep(options), { locals: scopedLocals })
+  const scopedOptions = merge(cloneDeep(data), { locals: scopedLocals })
   // walk through the contents and run replacements with modified options
   // we need to clone the node because the normal operation modifies
   // the node directly