--- a/lib/placeholders.js
+++ b/lib/placeholders.js
@@ -16,12 +16,12 @@ module.exports = function replace (input, ctx, settings) {
   // Since we are matching multiple sets of delimiters, we need to run a loop
   // here to match each one.
   for (let i = 0; i < settings.length; i++) {
-    const matchs = input.match(settings[i].regexp)
-    if (!matchs) continue
+    const matches = input.match(settings[i].regexp)
+    if (!matches) continue
 
     const delimiters = settings[i].text
-    for (let j = 0; j < matchs.length; j++) {
-      const match = matchs[j]
+    for (let j = 0; j < matches.length; j++) {
+      const match = matches[j]
       const expression = match.substring(delimiters[0].length, match.length - delimiters[1].length).trim()
 
       // If expression has non-word characters then use VM