--- a/test/fixtures/loop_conditional.html
+++ b/test/fixtures/loop_conditional.html
@@ -1,0 +1,12 @@
+<p>x</p>
+<ul>
+  <each loop='number in items'>
+    <if condition="number % 2 === 0">
+      <li>{{ number }} % 2 === 0</li>
+    </if>
+    <else>
+      <li>{{ number }}</li>
+    </else>
+  </each>
+</ul>
+<p>x</p>