## Proposed Changes

### Replacing the placeholders

`string.replace` 2-2.5x faster than `vm.runInContext`, so we can use `string.replace` for simple expression (without dynamic calculations).

If the expression has something besides letters and numbers - use **VM** otherwise **replace**.

#### Benchmark [150x `{{ key }}`]

| runInContext | runInContext + replace | faster |
|--------------|------------------------|--------|
| 234 op/s     | 530 op/s               | 2.3x   |

### Loop statement parser

Use new solution for parsing loop statement

#### Benchmark
  
| Current solution | New solution | faster |
|------------------|--------------|--------|
| 196 450 op/s      | 1 798 479 op/s | 9.2x   |

### Manual expansion the expression of conditions

#### Benchmark [20x `if - else if -else if - else`]

| Current solution | New solution | faster |
|------------------|--------------|--------|
| 322 op/s      | 349 op/s | 1.08x   |

## Types of Changes

Fully compatible with the current version.