@mrmlnc 2016-11-16
Don't merge this now, please. I want to add some tests.