@michael-ciniawsky 2016-11-16
@mrmlnc Whats your opinion on a `<switch>` tag, mimicking a switch statement?
```js
<switch>
  <case>
  <case>
  <case>
</switch>
```
likely this one is total rubbish 😛 , anyways...

`<js>Like a script tag but gets executed in Context</js>`