@mrmlnc 2016-11-16
@michael-ciniawsky,

I want to talk about it soon. Because all these changes do not solve the problem (a needle in a haystack), which was announced in #26.