@michael-ciniawsky 2016-11-16
Are there no known faster implementations of `merge` and `clonedeep` ? 😛