@mrmlnc 2016-11-16
@michael-ciniawsky,

Spoiler: even a simple `Object.assign({}, options)` kills performance from 500 op/s to 6-10 op/s.