@michael-ciniawsky 2016-11-16
uhhhh... 😟  I'm an extensive `Object.assign`  user