@mrmlnc 2016-11-16
@michael-ciniawsky,

Ready to merge. Please look again 😄