@michael-ciniawsky 2016-11-16
@mrmlnc I make few chore/docs tweaks later today and publish new under `posthtml-expressions`